addressed PR feedback for _add rest endpoint
